+++ examples/view_example.py
@@ -1,0 +1,13 @@
+from sanic import Sanic
+from sanic.response import text
+from examples import my_plugin
+
+app = Sanic(__name__)
+
+@app.route('/')
+@my_plugin.decorate(app)
+def index(request):
+    return text("hello world")
+
+if __name__ == "__main__":
+    app.run("127.0.0.1", port=8098, debug=True)