+++ spf/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: latin-1 -*-
 from .plugin import SanicPlugin, SanicPluginsFramework
-__version__ = '0.3.1.dev20171102'
+__version__ = '0.3.2.dev20171102'
 __all__ = ["SanicPlugin", "SanicPluginsFramework", "__version__"]